requirements.txt
torch==2.0.1
torchvision==0.15.2
Flask==2.1.1
numpy==1.21.6
matplotlib==3.5.1
seaborn==0.11.2
scikit-learn==1.0.2
Pillow==9.1.0